listar usuarios ya funciona

+++ tiendavirtual1/src/main/java/co/edu/unbosque/tiendavirtual1/dao/UsuariosDAO.java
@@ -1,8 +1,11 @@
 package co.edu.unbosque.tiendavirtual1.dao;
 
+import java.util.ArrayList;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import co.edu.unbosque.tiendavirtual1.model.Usuarios;
 
-public interface UsuariosDAO extends JpaRepository<Usuarios, Integer>{
+public interface UsuariosDAO extends JpaRepository<Usuarios, Long>{
 
 }
